Changed unsigned int to unsigned long ing

+++ inst/include/importExportMPZ.h
@@ -10,10 +10,10 @@ typedef void (*gmp_binary)(mpz_t, const mpz_t, const mpz_t);
  * Functions for importing/exporting and converting SEXPs to mpz_t
  * 
  */
-void createMPZArray (SEXP v, mpz_t myVec[], unsigned int sizevec);
+void createMPZArray (SEXP v, mpz_t myVec[], unsigned long int sizevec);
 
 int myRaw (char* raw, mpz_t value, unsigned long int totals);
 
-void quickSort(mpz_t arr[], int left, int right, std::vector<unsigned int>& lens);
+void quickSort(mpz_t arr[], int left, int right, std::vector<unsigned long int>& lens);
 
 #endif
